refactor: move d2 to vectors
