--- src/coordinax/_src/d2/__init__.py
@@ -1,18 +1,0 @@
-# pylint: disable=duplicate-code
-"""2-dimensional representations."""
-
-from . import base, cartesian, compat, polar, spherical, transform
-from .base import *
-from .cartesian import *
-from .compat import *
-from .polar import *
-from .spherical import *
-from .transform import *
-
-__all__: list[str] = []
-__all__ += base.__all__
-__all__ += cartesian.__all__
-__all__ += polar.__all__
-__all__ += spherical.__all__
-__all__ += transform.__all__
-__all__ += compat.__all__
