+++ src/coordinax/_src/transform/d2.py
@@ -10,9 +10,6 @@ from plum import dispatch
 import quaxed.numpy as jnp
 from unxt.quantity import AbstractQuantity, Quantity
 
-from coordinax._src.d2.base import AbstractPos2D
-from coordinax._src.d2.cartesian import CartesianPos2D
-from coordinax._src.d2.polar import PolarPos
 from coordinax._src.d3.base import AbstractPos3D
 from coordinax._src.d3.cartesian import CartesianPos3D
 from coordinax._src.d3.cylindrical import CylindricalPos
@@ -20,6 +17,7 @@ from coordinax._src.d3.mathspherical import MathSphericalPos
 from coordinax._src.d3.spherical import SphericalPos
 from coordinax._src.exceptions import IrreversibleDimensionChange
 from coordinax._src.vectors.d1 import CartesianPos1D, RadialPos
+from coordinax._src.vectors.d2 import AbstractPos2D, CartesianPos2D, PolarPos
 
 
 @dispatch.multi(
