+++ src/coordinax/_src/transform/d3.py
@@ -9,15 +9,13 @@ from plum import dispatch
 
 import quaxed.numpy as jnp
 
-from coordinax._src.d2.base import AbstractPos2D
-from coordinax._src.d2.cartesian import CartesianPos2D
-from coordinax._src.d2.polar import PolarPos
 from coordinax._src.d3.cartesian import CartesianPos3D
 from coordinax._src.d3.cylindrical import CylindricalPos
 from coordinax._src.d3.mathspherical import MathSphericalPos
 from coordinax._src.d3.spherical import SphericalPos
 from coordinax._src.exceptions import IrreversibleDimensionChange
 from coordinax._src.vectors.d1 import CartesianPos1D, RadialPos
+from coordinax._src.vectors.d2 import AbstractPos2D, CartesianPos2D, PolarPos
 
 # =============================================================================
 # CartesianPos3D
