+++ src/coordinax/_src/vectors/d2/base.py
