+++ src/coordinax/_src/vectors/d2/cartesian.py
