+++ src/coordinax/_src/vectors/d2/compat.py
